move asm_emitter
