+++ sharkfuser/include/fusilli.h
@@ -14,8 +14,8 @@
 #define FUSILLI_H
 
 // Support
-#include "fusilli/asm_emitter.h"
 #include "fusilli/extras.h"
+#include "fusilli/support/asm_emitter.h"
 #include "fusilli/support/external_tools.h"
 #include "fusilli/support/logging.h"
 #include "fusilli/types.h"
